♻️ 处理 Kafka 接入 review 问题
- AI handler 通过 context 复用 Kafka service，避免每次工具调用都重建 KafkaClientManager
- add_asset / update_asset 工具补全 Kafka 类型与 brokers/SASL/TLS 等参数
- app_kafka.go 与 app_redis.go 对齐，删除多余的 mutex 与悬空懒加载
- 抽取 connpool.BuildTLSConfig 公共函数，复用 Redis/Kafka 的 TLS 配置构建
- 将 ArgBool / ArgStringSlice 上移到 assettype/registry.go 供其他类型复用
- 拆分 kafka_helper.go 为 kafka_helper / kafka_command / kafka_args 三个文件
