+++ internal/ai/tool_registry.go
@@ -83,14 +83,14 @@ func AllToolDefs() []ToolDef {
 		},
 		{
 			Name:        "add_asset",
-			Description: `Add a new asset to the inventory. Supports types: "ssh", "database", "redis", "mongodb", "k8s". For database, specify driver ("mysql" or "postgresql"). For k8s, specify kubeconfig. Credentials (password / private_key) are stored encrypted; never echo them back to the user.`,
+			Description: `Add a new asset to the inventory. Supports types: "ssh", "database", "redis", "mongodb", "kafka", "k8s". For database, specify driver ("mysql" or "postgresql"). For k8s, specify kubeconfig. For kafka, specify brokers (comma-separated) plus optional sasl_mechanism / tls. Credentials (password / private_key) are stored encrypted; never echo them back to the user.`,
 			Params: []ParamDef{
 				{Name: "name", Type: ParamString, Description: `Display name for the asset.`, Required: true},
-				{Name: "type", Type: ParamString, Description: `Asset type: "ssh" (default), "database", "redis", "mongodb", or "k8s".`},
-				{Name: "host", Type: ParamString, Description: "Hostname or IP address. Required except for k8s."},
-				{Name: "port", Type: ParamNumber, Description: "Port number (default: 22 for SSH, 3306 for MySQL, 5432 for PostgreSQL, 6379 for Redis, 27017 for MongoDB). Required except for k8s."},
-				{Name: "username", Type: ParamString, Description: "Login username. Required except for k8s."},
-				{Name: "password", Type: ParamString, Description: "Plaintext password. Stored encrypted by the app. For SSH password auth, or database/redis/mongodb."},
+				{Name: "type", Type: ParamString, Description: `Asset type: "ssh" (default), "database", "redis", "mongodb", "kafka", or "k8s".`},
+				{Name: "host", Type: ParamString, Description: "Hostname or IP address. Required except for k8s and kafka (kafka uses brokers, but host+port falls back to a single broker)."},
+				{Name: "port", Type: ParamNumber, Description: "Port number (default: 22 for SSH, 3306 for MySQL, 5432 for PostgreSQL, 6379 for Redis, 27017 for MongoDB, 9092 for Kafka). Required except for k8s; for kafka, used only when brokers is omitted."},
+				{Name: "username", Type: ParamString, Description: "Login username. Required for ssh/database/redis/mongodb; optional for kafka (only when SASL is configured)."},
+				{Name: "password", Type: ParamString, Description: "Plaintext password. Stored encrypted by the app. For SSH password auth, database/redis/mongodb, or kafka SASL auth."},
 				{Name: "auth_type", Type: ParamString, Description: `SSH auth method: "password" (default if password supplied) or "key" (default if private_key supplied). Only for SSH type.`},
 				{Name: "private_key", Type: ParamString, Description: "SSH private key in PEM format. Imported into the credential store and linked to the asset. SSH only."},
 				{Name: "passphrase", Type: ParamString, Description: "Passphrase for the SSH private key, if encrypted. SSH only."},
@@ -101,7 +101,12 @@ func AllToolDefs() []ToolDef {
 				{Name: "kubeconfig", Type: ParamString, Description: "Kubeconfig YAML content. Required for k8s type."},
 				{Name: "namespace", Type: ParamString, Description: "Default Kubernetes namespace. K8s only."},
 				{Name: "context", Type: ParamString, Description: "Kubeconfig context name to use. K8s only."},
-				{Name: "ssh_asset_id", Type: ParamNumber, Description: "SSH asset ID for tunnel connection. For database / redis / mongodb / k8s types."},
+				{Name: "brokers", Type: ParamString, Description: `Kafka broker list as comma/semicolon/newline separated "host:port" pairs (e.g. "kafka-0:9092,kafka-1:9092"). Kafka only; falls back to host+port when omitted.`},
+				{Name: "client_id", Type: ParamString, Description: "Kafka client ID label. Kafka only; defaults to opskat."},
+				{Name: "sasl_mechanism", Type: ParamString, Description: `Kafka SASL mechanism: "none" (default), "plain", "scram-sha-256", "scram-sha-512". Kafka only.`},
+				{Name: "tls", Type: ParamString, Description: `Set to "true" to enable TLS. Kafka only.`},
+				{Name: "tls_insecure", Type: ParamString, Description: `Set to "true" to skip TLS certificate verification. Kafka only.`},
+				{Name: "ssh_asset_id", Type: ParamNumber, Description: "SSH asset ID for tunnel connection. For database / redis / mongodb / kafka / k8s types."},
 				{Name: "group_id", Type: ParamNumber, Description: "Group ID to assign this asset to."},
 				{Name: "description", Type: ParamString, Description: "Optional description or notes."},
 			},
@@ -124,7 +129,12 @@ func AllToolDefs() []ToolDef {
 				{Name: "database", Type: ParamString, Description: "New default database. Database / mongodb only. Pass empty string to clear."},
 				{Name: "read_only", Type: ParamString, Description: `Set to "true"/"false" to toggle read-only. Database only.`},
 				{Name: "redis_db", Type: ParamNumber, Description: "New default Redis DB index. Redis only."},
-				{Name: "ssh_asset_id", Type: ParamNumber, Description: "New SSH tunnel asset ID. Pass 0 to detach. Database / redis / mongodb only."},
+				{Name: "brokers", Type: ParamString, Description: `New Kafka broker list (comma/semicolon/newline separated "host:port"). Kafka only.`},
+				{Name: "client_id", Type: ParamString, Description: "New Kafka client ID. Kafka only."},
+				{Name: "sasl_mechanism", Type: ParamString, Description: `New Kafka SASL mechanism: "none", "plain", "scram-sha-256", "scram-sha-512". Kafka only.`},
+				{Name: "tls", Type: ParamString, Description: `Set to "true"/"false" to toggle Kafka TLS. Kafka only.`},
+				{Name: "tls_insecure", Type: ParamString, Description: `Set to "true"/"false" to toggle Kafka TLS skip-verify. Kafka only.`},
+				{Name: "ssh_asset_id", Type: ParamNumber, Description: "New SSH tunnel asset ID. Pass 0 to detach. Database / redis / mongodb / kafka only."},
 				{Name: "description", Type: ParamString, Description: "New description. Pass empty string to clear."},
 				{Name: "group_id", Type: ParamNumber, Description: "New group ID (must be a positive integer from list_groups). Omit to keep current group; values <= 0 are ignored. To remove an asset from its group, ask the user to do it in the UI."},
 				{Name: "icon", Type: ParamString, Description: "New icon name."},
