+++ internal/app/app.go
@@ -77,7 +77,6 @@ type App struct {
 	sshProxyServer          *sshpool.Server            // SSH 连接池 Unix socket 服务
 	redisService            *redis_svc.Service         // Redis 浏览/编辑服务
 	kafkaService            *kafka_svc.Service         // Kafka 管理服务
-	kafkaServiceMu          sync.Mutex                 // 保护 Kafka service 懒初始化
 	shutdownCh              chan struct{}              // 关闭信号，cleanup 时 close 以解除所有阻塞等待
 	pendingAuthResponses    sync.Map                   // map[string]chan []string（keyboard-interactive 认证响应用）
 	pendingHostKeyResponses sync.Map                   // map[string]chan ssh_svc.HostKeyAction（主机密钥校验响应用）
@@ -122,9 +121,7 @@ func (a *App) Startup(ctx context.Context) {
 	a.startApprovalServer(authToken)
 	a.startSSHPoolServer(authToken)
 	a.redisService = redis_svc.New(a.sshPool)
-	a.kafkaServiceMu.Lock()
 	a.kafkaService = kafka_svc.New(a.sshPool)
-	a.kafkaServiceMu.Unlock()
 	a.startAutoUpdateCheck()
 	a.InitAIProvider()
 	a.subscribeAIFlushAck()
@@ -193,12 +190,10 @@ func (a *App) Cleanup() {
 	// 先发送关闭信号，解除所有阻塞等待（审批、权限确认等），避免 wg.Wait 死锁
 	close(a.shutdownCh)
 
-	a.kafkaServiceMu.Lock()
 	if a.kafkaService != nil {
 		a.kafkaService.Close()
 		a.kafkaService = nil
 	}
-	a.kafkaServiceMu.Unlock()
 	if a.sshProxyServer != nil {
 		a.sshProxyServer.Stop()
 	}
