+++ internal/app/app_kafka.go
@@ -9,8 +9,6 @@ import (
 )
 
 func (a *App) kafkaSvc() *kafka_svc.Service {
-	a.kafkaServiceMu.Lock()
-	defer a.kafkaServiceMu.Unlock()
 	if a.kafkaService == nil {
 		a.kafkaService = kafka_svc.New(a.sshPool)
 	}
