+++ internal/assettype/registry.go
@@ -4,6 +4,7 @@ package assettype
 import (
 	"context"
 	"fmt"
+	"strings"
 	"sync"
 
 	"github.com/opskat/opskat/internal/model/entity/asset_entity"
@@ -105,4 +106,55 @@ func ArgInt64(args map[string]any, key string) int64 {
 	default:
 		return 0
 	}
+}
+
+// ArgBool 从 args 中解析 bool。支持 bool、字符串 ("true"/"1"/"yes")、数字 1。
+func ArgBool(args map[string]any, key string) bool {
+	v, ok := args[key]
+	if !ok {
+		return false
+	}
+	switch x := v.(type) {
+	case bool:
+		return x
+	case string:
+		return strings.EqualFold(x, "true") || x == "1" || strings.EqualFold(x, "yes")
+	default:
+		return fmt.Sprintf("%v", x) == "1"
+	}
+}
+
+// ArgStringSlice 从 args 中解析字符串数组。支持 []string、[]any、用逗号/分号/换行分隔的字符串。
+// 自动 trim 空白并丢弃空项。
+func ArgStringSlice(args map[string]any, key string) []string {
+	v, ok := args[key]
+	if !ok || v == nil {
+		return nil
+	}
+	switch x := v.(type) {
+	case []string:
+		return cleanStrings(x)
+	case []any:
+		out := make([]string, 0, len(x))
+		for _, item := range x {
+			out = append(out, fmt.Sprintf("%v", item))
+		}
+		return cleanStrings(out)
+	case string:
+		parts := strings.FieldsFunc(x, func(r rune) bool { return r == ',' || r == '\n' || r == ';' })
+		return cleanStrings(parts)
+	default:
+		return nil
+	}
+}
+
+func cleanStrings(values []string) []string {
+	out := make([]string, 0, len(values))
+	for _, v := range values {
+		v = strings.TrimSpace(v)
+		if v != "" {
+			out = append(out, v)
+		}
+	}
+	return out
 }
