+++ internal/connpool/kafka.go
@@ -3,11 +3,9 @@ package connpool
 import (
 	"context"
 	"crypto/tls"
-	"crypto/x509"
 	"fmt"
 	"hash/fnv"
 	"net"
-	"os"
 	"sort"
 	"strconv"
 	"strings"
@@ -344,33 +342,13 @@ func splitKafkaAddr(addr string) (string, int, error) {
 }
 
 func buildKafkaTLSConfig(cfg *asset_entity.KafkaConfig) (*tls.Config, error) {
-	tlsConfig := &tls.Config{
-		MinVersion:         tls.VersionTLS12,
-		ServerName:         cfg.TLSServerName,
-		InsecureSkipVerify: cfg.TLSInsecure,
-	}
-	if cfg.TLSCAFile != "" {
-		ca, err := os.ReadFile(cfg.TLSCAFile)
-		if err != nil {
-			return nil, fmt.Errorf("读取 Kafka TLS CA 证书失败: %w", err)
-		}
-		pool := x509.NewCertPool()
-		if !pool.AppendCertsFromPEM(ca) {
-			return nil, fmt.Errorf("解析 Kafka TLS CA 证书失败")
-		}
-		tlsConfig.RootCAs = pool
-	}
-	if cfg.TLSCertFile != "" || cfg.TLSKeyFile != "" {
-		if cfg.TLSCertFile == "" || cfg.TLSKeyFile == "" {
-			return nil, fmt.Errorf("kafka TLS 客户端证书和私钥必须同时配置")
-		}
-		cert, err := tls.LoadX509KeyPair(cfg.TLSCertFile, cfg.TLSKeyFile)
-		if err != nil {
-			return nil, fmt.Errorf("加载 Kafka TLS 客户端证书失败: %w", err)
-		}
-		tlsConfig.Certificates = []tls.Certificate{cert}
-	}
-	return tlsConfig, nil
+	return BuildTLSConfig("Kafka", TLSFields{
+		ServerName: cfg.TLSServerName,
+		Insecure:   cfg.TLSInsecure,
+		CAFile:     cfg.TLSCAFile,
+		CertFile:   cfg.TLSCertFile,
+		KeyFile:    cfg.TLSKeyFile,
+	})
 }
 
 func hashString(value string) string {
